syndwarsfx: Remove the non-original quit shortcut
In the first port, Ctrl+Q was added as hard-coded quit
keyboard shortcut. Disable it, at it leads to mistakes
while trying to use cheats.

Also, we will soon switch to cheats using only keys
with modifiers, and Ctrl might be a good modifier to replace
the "just q" cheat.