syndwarsfx: Introduced level strain and variant
Looks like the idea was that single level number is there
only to condense storage, and really the game assumes
levelh have two numbers: a strain / idea for a level,
and variant/version within that strain.

There can be 15 strains for each map, currently.
Since level number is sometimes stored as byte,
variants should only be used up to 15 as well.