syndwarsfx: Implement game_set_cam_starting_location_xz()
Starting location, when loading a level, was set to specific
point if mounted guns are on level, and retained from whatever
was stored previously otherwise. The position would also
sometimes reset on restart, and sometimes be kept.

Now the rules are more deliberate. If mounted guns exist on map
they are still used, as original levels with custom camera
movement expect them in these positions, changing this rule
would make the camera fucus on uninteresting things.

If no mounted gun is found, position of a billboard is taken.

If no billboard as well, position of opposite part of the map than
location of the first agent.

If a level is restarted, position at which 'R' key was pressed is
always retained. This introduced interesting effect as the player
sees the same area, just with time reversed to the beginning of
the mission.